renamed FieldDescription -> Description
